backend/pyproject.toml
[project]
name = "budongsan-law-backend"
version = "0.1.0"
description = "FastAPI backend for the Hybrid Legal Q&A platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.2",
  "uvicorn[standard]==0.30.6",
  "pydantic-settings==2.5.2",
  "httpx==0.27.2",
  "PyJWT[crypto]==2.9.0",
  "python-multipart==0.0.10",
]

[project.optional-dependencies]
dev = [
  "ruff==0.6.9",
]

[tool.uv]
pyproject-support = true
